Update README.md
Clarified the set up process for Windows users, since simply copy and pasting the echo command in Running Locally will result in invalid json on windows. I also added a sentence to the Building Binaries section stating that this step is unnecessary if you just want to run the game.